Etapa 1 (Fundação): pricing_configs versionada, créditos fracionados
(Numeric 12,4) nos saldos e ledgers, tipos resale/adjustment e auditoria
de tokens input/output.

Etapa 2 (Consumo): cálculo ponderado de créditos via pricing_configs
(substitui o ceil fixo por tokens), moeda única no consumo do cliente
final (debita só a wallet dele quando custeado por ele, nunca os dois),
lock de concorrência no débito (SELECT ... FOR UPDATE), migration da
política de saldo insuficiente (hook) e saldos/valores de crédito como
float na API com formatação pt-BR no painel.

+++ apps/api/app/models/tenant.py
@@ -1,7 +1,8 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import CheckConstraint, DateTime, Integer, String, Uuid, text
+from sqlalchemy import CheckConstraint, DateTime, Numeric, String, Uuid, text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base
@@ -22,7 +23,9 @@ class Tenant(Base):
     logo_filename: Mapped[str | None] = mapped_column(String)
     # Cache do saldo — fonte da verdade é o ledger em credit_transactions,
     # atualizado na mesma transação de cada lançamento.
-    credit_balance: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text("0"))
+    credit_balance: Mapped[Decimal] = mapped_column(
+        Numeric(12, 4), nullable=False, server_default=text("0")
+    )
     status: Mapped[str] = mapped_column(String, nullable=False, server_default=text("'active'"))
     onboarding_completed_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     created_at: Mapped[datetime] = mapped_column(
